chore: Fragment 제거
